// src/comments/api/dto/output.comment.dto.ts
import { LikeStatus } from '../../../global-types/global.types';

export class OutputCommentDto {
  id: string;
  content: string;
  commentatorInfo: {
    userId: string;
    userLogin: string;
  };
  createdAt: string;
  likesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: LikeStatus;
  };
}